requirements.txt
flask>=2.0.0
google-auth>=2.0.0
PyJWT>=2.0.0
google-cloud-storage>=2.0.0
requests>=2.26.0
python-dotenv>=0.19.0